+++ demo/demo.html
@@ -8,16 +8,17 @@
 
 <body>
 
-	<div class="demo1" data-init="(function($elm) { $elm.plugin(); })" data-resources="[
+	<div class="wrapper">
+		<div class="demo1" data-init="(function($elm) { $elm.plugin(); })" data-resources="[
 			 	{
 			 		test: (function() {return true;}),
-					 paths: ['css/demo.css', 'css/demo1.css', 'https://code.jquery.com/jquery-3.3.1.min.js']
+					 paths: ['css/demo.css?speed=20', 'css/demo1.css', 'https://code.jquery.com/jquery-3.3.1.min.js']
 			 	}
 			 ]">
-		This component's resources (css and js) are loaded via conditional loader!
-	</div>
+			This component's resources (css and js) are loaded via conditional loader!
+		</div>
 
-	<div class="demo2" data-init="(function($elm) { $elm.plugin(); })" data-resources="[
+		<div class="demo2" data-init="(function($elm) { $elm.plugin(); })" data-resources="[
 			 	{
 					 paths: ['/demo/resources/js/jquery.plugin.js', '/demo/resources/css/demo.css?speed=20'],
 					 dependsOn: ['https://code.jquery.com/jquery-3.3.1.min.js']
@@ -27,10 +28,10 @@
 			 		base: '##content'
 			 	}
 			 ]">
-		This component's resources are loaded via conditional loader, using relative and absolute paths.
-	</div>
+			This component's resources are loaded via conditional loader, using relative and absolute paths.
+		</div>
 
-	<div class="demo3" data-init="(function($elm) { $elm.plugin(); $elm.component(); })" data-resources="[
+		<div class="demo3" data-init="(function($elm) { $elm.plugin(); $elm.component(); })" data-resources="[
 			 	{
 					 paths: ['/demo/resources/js/jquery.plugin.js', '/demo/resources/css/demo3.css'],
 					 dependsOn: ['https://code.jquery.com/jquery-3.3.1.min.js']
@@ -40,18 +41,14 @@
 			 		dependsOn: ['js/jquery.plugin.js', 'https://code.jquery.com/jquery-3.3.1.min.js']
 			 	}
 			 ]">
-		This component's resources are loaded via conditional loader. One resource depends on another. plugin.js -> component.js
-	</div>
+			This component's resources are loaded via conditional loader. One resource depends on another. plugin.js -> component.js
+		</div>
 
-	<div class="demo4" data-init="(function($elm) { $elm.component2(); $elm.component(); $elm.component3(); })" data-resources="
+		<div class="demo4" data-init="(function($elm) { $elm.component2(); $elm.component(); $elm.component3(); })" data-resources="
 			 [
 				{
-					paths: ['js/jquery.component.js', 'css/demo.css'],
+					paths: ['js/jquery.component.js', 'css/demo.css?speed=20'],
 					dependsOn: ['https://code.jquery.com/jquery-3.3.1.min.js']
-				},
-				{
-					paths: ['css/demo4.css'],
-					base: '##content'
 				},
 			 	{
 			 		paths: ['js/jquery.component2.js'],
@@ -61,13 +58,31 @@
 			 		dependsOn: ['js/jquery.component2.js', '/demo/resources/js/jquery.plugin.js', 'https://code.jquery.com/jquery-3.3.1.min.js']
 			 	}
 			 ]">
-		Resources are loaded in order of dependencies: component.js -> component2.js -> component3.js
-		<br/> plugin.js is loaded on other demos or skipped. if skipped an error will be thrown, but component3.js will still be loaded
+			Resources are loaded in order of dependencies: component.js -> component2.js -> component3.js
+			<br/> plugin.js is loaded on other demos or skipped. if skipped an error will be thrown, but component3.js will still be loaded
+		</div>
+		<div class="demo4" data-init="(function($elm) { $elm.component2(); $elm.component(); $elm.component3(); })" data-resources="
+			 [
+				{
+					paths: ['js/jquery.component.js', 'css/demo.css?speed=20'],
+					dependsOn: ['https://code.jquery.com/jquery-3.3.1.min.js']
+				},
+			 	{
+			 		paths: ['js/jquery.component2.js'],
+			 		dependsOn: ['js/jquery.component.js', 'https://code.jquery.com/jquery-3.3.1.min.js']
+			 	}, {
+			 		paths: ['js/jquery.component3.js', 'js/jquery.plugin.js'],
+			 		dependsOn: ['js/jquery.component2.js', '/demo/resources/js/jquery.plugin.js', 'https://code.jquery.com/jquery-3.3.1.min.js']
+			 	}
+			 ]">
+			Resources are loaded in order of dependencies: component.js -> component2.js -> component3.js
+			<br/> plugin.js is loaded on other demos or skipped. if skipped an error will be thrown, but component3.js will still be loaded
+		</div>
 	</div>
 	<div class="demo5" data-init="(function($elm) { $elm.component2(); $elm.component(); $elm.component3(); })" data-resources="
 			 [
 				{
-					paths: ['js/jquery.component.js', 'css/demo.css'],
+					paths: ['js/jquery.component.js'],
 					dependsOn: ['https://code.jquery.com/jquery-3.3.1.min.js']
 				},
 				{
@@ -85,14 +100,22 @@
 		Resources are loaded in order of dependencies: component.js -> component2.js -> component3.js
 		<br/> plugin.js is loaded on other demos or skipped. if skipped an error will be thrown, but component3.js will still be loaded
 	</div>
+	<button>Load new again 👏</button>
+	<!-- <script src="https://code.jquery.com/jquery-3.3.1.min.js"></script> -->
 	<script src="../dist/resourceLoader.min.js"></script>
 	<script>
-		resourceLoader({
+		resourceLoader.init({
 			base: 'resources/',
+			container: '.wrapper',
 			baseMap: {
 				'##content': 'resources-content/'
 			}
 		});
+		const button = document.querySelector('button')
+		button.addEventListener('click', () => {
+			const container = document.querySelector('body');
+			resourceLoader.update(container);
+		})
 	</script>
 </body>
