--- demo/oldDemo.html
@@ -1,100 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-  <meta charset="utf-8" />
-  <title>Demo for conditional loader</title>
-</head>
-<script src="http://code.jquery.com/jquery-3.1.1.js"></script>
-
-<body>
-
-  <div class="demo1" data-init="(function($elm) { $elm.plugin(); })" data-resources="[
-			 	{
-			 		test: (function() {return true;}),
-					 paths: ['css/demo.css', 'css/demo1.css', 'https://code.jquery.com/jquery-3.3.1.min.js']
-			 	}
-			 ]">
-    This component's resources (css and js) are loaded via conditional loader!
-  </div>
-
-  <div class="demo2" data-init="(function($elm) { $elm.plugin(); })" data-resources="[
-			 	{
-					 paths: ['/demo/resources/js/jquery.plugin.js', '/demo/resources/css/demo.css?speed=20'],
-					 dependsOn: ['https://code.jquery.com/jquery-3.3.1.min.js']
-			 	},
-			 	{
-			 		paths: ['css/demo2.css'],
-			 		base: '##content'
-			 	}
-			 ]">
-    This component's resources are loaded via conditional loader, using relative and absolute paths.
-  </div>
-
-  <div class="demo3" data-init="(function($elm) { $elm.plugin(); $elm.component(); })" data-resources="[
-			 	{
-					 paths: ['/demo/resources/js/jquery.plugin.js', '/demo/resources/css/demo3.css'],
-					 dependsOn: ['https://code.jquery.com/jquery-3.3.1.min.js']
-			 	},
-			 	{
-			 		paths: ['js/jquery.component.js'],
-			 		dependsOn: ['js/jquery.plugin.js', 'https://code.jquery.com/jquery-3.3.1.min.js']
-			 	}
-			 ]">
-    This component's resources are loaded via conditional loader. One resource depends on another. plugin.js -> component.js
-  </div>
-
-  <div class="demo4" data-init="(function($elm) { $elm.component2(); $elm.component(); $elm.component3(); })" data-resources="
-			 [
-				{
-					paths: ['js/jquery.component.js', 'css/demo.css'],
-					dependsOn: ['https://code.jquery.com/jquery-3.3.1.min.js']
-				},
-				{
-					paths: ['css/demo4.css'],
-					base: '##content'
-				},
-			 	{
-			 		paths: ['js/jquery.component2.js'],
-			 		dependsOn: ['js/jquery.component.js', 'https://code.jquery.com/jquery-3.3.1.min.js']
-			 	}, {
-			 		paths: ['js/jquery.component3.js', 'js/jquery.plugin.js'],
-			 		dependsOn: ['js/jquery.component2.js', '/demo/resources/js/jquery.plugin.js', 'https://code.jquery.com/jquery-3.3.1.min.js']
-			 	}
-			 ]">
-    Resources are loaded in order of dependencies: component.js -> component2.js -> component3.js
-    <br/> plugin.js is loaded on other demos or skipped. if skipped an error will be thrown, but component3.js will still be loaded
-  </div>
-  <div class="demo5" data-init="(function($elm) { $elm.component2(); $elm.component(); $elm.component3(); })" data-resources="
-			 [
-				{
-					paths: ['js/jquery.component.js', 'css/demo.css'],
-					dependsOn: ['https://code.jquery.com/jquery-3.3.1.min.js']
-				},
-				{
-					paths: ['css/demo4.css'],
-					base: '##content'
-				},
-			 	{
-			 		paths: ['js/jquery.component2.js'],
-			 		dependsOn: ['js/jquery.component.js', 'https://code.jquery.com/jquery-3.3.1.min.js']
-			 	}, {
-			 		paths: ['js/jquery.component3.js', 'js/jquery.plugin.js'],
-			 		dependsOn: ['js/jquery.component2.js', '/demo/resources/js/jquery.plugin.js', 'https://code.jquery.com/jquery-3.3.1.min.js']
-			 	}
-			 ]">
-    Resources are loaded in order of dependencies: component.js -> component2.js -> component3.js
-    <br/> plugin.js is loaded on other demos or skipped. if skipped an error will be thrown, but component3.js will still be loaded
-  </div>
-  <script src="../dist/oldresourceLoader.js"></script>
-  <script>
-    resourceLoader({
-      base: 'resources/',
-      baseMap: {
-        '##content': 'resources-content/'
-      }
-    });
-  </script>
-</body>
-
-</html>
