+++ demo/resources/js/jquery.component.js
@@ -9,14 +9,13 @@
  * based on: http://jqueryboilerplate.com/
  */
 
-;(function ($, window, document, undefined) {
+(function($, window, document, undefined) {
   'use strict';
 
-  var pluginName = 'component',
+  let pluginName = 'component',
     defaults = {
       foo: 'bar'
     };
-
   // The actual plugin constructor
   function Plugin(element, options) {
     this.$element = $(element);
@@ -25,13 +24,13 @@
   }
 
   // methods
-  var methods = {
+  const methods = {
     init: function() {
       // your init is goes here
       this.myFunction();
     },
 
-    myFunction: function(){
+    myFunction: function() {
       console.log('initializing component');
     }
   };
@@ -43,23 +42,22 @@
   // preventing against multiple instantiations
   $.fn[pluginName] = function(options) {
     this.each(function() {
-      if(!$.data(this, 'plugin_' + pluginName)) {
+      if (!$.data(this, 'plugin_' + pluginName)) {
         $.data(this, 'plugin_' + pluginName, new Plugin(this, options));
 
         /**
-         * DEV-note: Use trigger failsafe events if other plugins depend on this plugins ready state
-         */
+				 * DEV-note: Use trigger failsafe events if other plugins depend on this plugins ready state
+				 */
         // trigger ready event on element - optional
-        //$.triggerFailsafeEvent($(this), 'plugin_' + pluginName + '.ready');
+        // $.triggerFailsafeEvent($(this), 'plugin_' + pluginName + '.ready');
       }
     });
 
     // trigger ready event globally - optional
-    //if(this.length > 0) {
+    // if(this.length > 0) {
     //    $.triggerFailsafeEvent($(window), 'plugin_' + pluginName + '.ready_all');
-    //}
+    // }
 
     return this;
   };
-
 })(jQuery, window, document);
