--- demo/resources/js/jquery.smoke.js
@@ -1,1 +1,0 @@
-$('body').html('mega');
