+++ package.json
@@ -1,16 +1,16 @@
 {
   "name": "@biotope/resource-loader",
   "version": "1.3.0",
-  "description": "a markup driven conditional loader to load resources (js and css) with dependencies",
+  "description":
+    "a markup driven conditional loader to load resources (js and css) with dependencies",
   "main": "src/resourceLoader.js",
   "repository": "git@github.com:biotope/biotope-resource-loader.git",
   "author": "Biotope",
   "license": "MIT",
   "scripts": {
     "build": "rollup -c",
-    "demo": "python -m SimpleHTTPServer 8000",
-    "dev": "rollup -c -w",
-    "speed": "./speed"
+    "serve": "./bin/speed",
+    "dev": "rollup -c -w"
   },
   "dependencies": {
     "deep-equal": "^1.0.1",
