--- test/css.html
@@ -1,28 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <meta http-equiv="X-UA-Compatible" content="ie=edge">
-  <title>Document</title>
-</head>
-
-<body>
-  <div class="wrapper">
-    <div class="demo1">hallo</div>
-    <div class="demo2">du</div>
-    <div class="demo3">held</div>
-  </div>
-  <script>
-    const myCSS = document.createElement("link");
-    myCSS.rel = "stylesheet";
-    myCSS.href = "http://localhost:8080/demo/resources/css/demo.css?speed=20";
-    document.head.insertBefore(myCSS, document.head.childNodes[document.head.childNodes.length - 1].nextSibling);
-    myCSS.addEventListener('load', () => {
-      console.log('we are ready')
-    })
-  </script>
-</body>
-
-</html>
