--- test/main.js
@@ -1,1 +1,0 @@
-alert('test');
